React Page modification CRUD

+++ back_end/index.js
@@ -6,6 +6,7 @@ const cors = require("cors");
 const sharp = require('sharp');
 const port = 5000;
 const fs = require("fs");
+const path = require('path')
 const imageModel = require("./models/image");
 require("./connection/connection")
 app.use(cors());
@@ -50,7 +51,7 @@ app.post('/api/images', upload.single('testImage'), generateThumbnail, async (re
     try {
         //const { originalname, size, path } = req.file;
         const originalname = req.file.filename;
-        const size = req.file.size;
+        const size = req.body.size;
         const path = req.file.path;
 
         //const { name, image_details, tags, thumbnail } = req.body;
@@ -101,7 +102,7 @@ app.get("/image/:id", async (req, res) => {
     try {
         const _id = req.params.id;
         const data = await imageModel.findById({ _id });
-        res.json(data);
+        res.status(200).json({ data: { ...data._doc } });
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: "Failed to fetch image data." });
@@ -111,6 +112,19 @@ app.get("/image/:id", async (req, res) => {
 app.delete("/image/:id", async (req, res) => {
     try {
         const _id = req.params.id;
+        const data = await imageModel.findById(_id);
+        const imagePath = path.join(__dirname, 'uploads', data.name);
+        const imageThumbnail = path.join(__dirname, data.thumbnail)
+        if (!fs.existsSync(imagePath)) {
+            return res.status(404).send("Image not found");
+        }
+        fs.unlinkSync(imagePath);
+
+        if (!fs.existsSync(imageThumbnail)) {
+            return res.status(404).send("Image not found");
+        }
+        fs.unlinkSync(imageThumbnail);
+
         const deleteData = await imageModel.findByIdAndDelete(_id);
         if (!_id) {
             return res.status(400).send()
